# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/register":{"post":{"summary":"Register User","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"summary":"Login User","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_user_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"summary":"Logout User","operationId":"logout_user_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/forms/create":{"post":{"summary":"Create Form","operationId":"create_form_forms_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FormCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Form"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"LoginManager":[]}]}},"/forms/delete/{form_id}":{"delete":{"summary":"Delete Form","operationId":"delete_form_forms_delete__form_id__delete","security":[{"LoginManager":[]}],"parameters":[{"name":"form_id","in":"path","required":true,"schema":{"type":"integer","title":"Form Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Form"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/forms/":{"get":{"summary":"Get All Forms","operationId":"get_all_forms_forms__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Form"},"type":"array","title":"Response Get All Forms Forms  Get"}}}}}}},"/forms/{form_id}":{"get":{"summary":"Get Single Form","operationId":"get_single_form_forms__form_id__get","parameters":[{"name":"form_id","in":"path","required":true,"schema":{"type":"integer","title":"Form Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Form"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/forms/submit/{form_id}":{"post":{"summary":"Submit Form","operationId":"submit_form_forms_submit__form_id__post","parameters":[{"name":"form_id","in":"path","required":true,"schema":{"type":"integer","title":"Form Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FormSubmissionCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/forms/submissions/{form_id}":{"get":{"summary":"Get Form Submissions","operationId":"get_form_submissions_forms_submissions__form_id__get","parameters":[{"name":"form_id","in":"path","required":true,"schema":{"type":"integer","title":"Form Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedSubmissions"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_user_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_user_auth_login_post"},"Field":{"properties":{"field_id":{"type":"string","title":"Field Id"},"type":{"type":"string","title":"Type"},"label":{"type":"string","title":"Label"},"required":{"type":"boolean","title":"Required"}},"type":"object","required":["field_id","type","label","required"],"title":"Field"},"Form":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"owner_id":{"type":"integer","title":"Owner Id"}},"type":"object","required":["title","description","id","created_at","owner_id"],"title":"Form"},"FormCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"fields":{"items":{"$ref":"#/components/schemas/Field"},"type":"array","title":"Fields"}},"type":"object","required":["title","description","fields"],"title":"FormCreate"},"FormSubmissionCreate":{"properties":{"responses":{"items":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"boolean"}]},"type":"object"},"type":"array","title":"Responses"}},"type":"object","required":["responses"],"title":"FormSubmissionCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaginatedSubmissions":{"properties":{"total_count":{"type":"integer","title":"Total Count"},"page":{"type":"integer","title":"Page"},"limit":{"type":"integer","title":"Limit"},"submissions":{"items":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"boolean"}]},"type":"object"},"type":"array","title":"Submissions"}},"type":"object","required":["total_count","page","limit","submissions"],"title":"PaginatedSubmissions"},"User":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["email","id","name"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"name":{"type":"string","title":"Name"}},"type":"object","required":["email","password","name"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"LoginManager":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}